requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
django-tinymce==3.5.0
Pillow==9.4.0
sqlparse==0.4.3
geoip2==4.6.0
django-comments-dab==2.8.0
crispy-bootstrap4==2022.1